+++ willow/plugins/opencv/__init__.py
@@ -1,8 +1,8 @@
 try:
-    from .opencv2 import OpenCVColorImage, OpenCVGrayscaleImage
+    from .opencv3 import OpenCVColorImage, OpenCVGrayscaleImage
 except ImportError:
     try:
-        from .opencv1 import OpenCVColorImage, OpenCVGrayscaleImage
+        from .opencv2 import OpenCVColorImage, OpenCVGrayscaleImage
     except ImportError:
         # Import the base class to access the check method for the registry.
         from .base import BaseOpenCVColorImage as OpenCVColorImage
